requirements.txt
# Flask and extensions
flask==2.0.1
flask-cors==3.0.10
python-dotenv==0.19.0

# Database
prisma==0.8.2

# OAuth and API
requests==2.26.0
python-jose==3.3.0

# Notion SDK
notion-client==1.0.0

# DocuSign SDK
docusign-esign==3.20.0

# Utils
pydantic==1.8.2 